0.3.5 Bot - Refactor #2
