+++ Telegram Bot/botfunctions/change_profile.py
@@ -1,3 +1,4 @@
+#Package to store profile change functions
 import mysql.connector
 from aiogram.utils import executor
 from aiogram.types import ParseMode
@@ -32,28 +33,11 @@ PARENT_FOLDER_ID = "1wb4h1vSTqsXxYB3-ah_r4cREMQc1ySPR"
 
 # Database connection, we will use mySQL and localhost for now
 
-pwd = None
-with open("includes\database_pwd.txt") as f:
-    pwd = f.read().strip()
-db = mysql.connector.connect(
-    host="localhost",
-    user='root',
-    passwd=pwd,
-    database="testdatabase"
-)
-
-mycursor = db.cursor()
+from botfunctions.databaseconn_dispatcher import db, dp
 
 logging.basicConfig(level=logging.INFO)
 
-# Set up dispatcher
-TOKEN = None
-with open("includes\dot_token.txt") as f:
-    TOKEN = f.read().strip()
 
-bot = Bot(token=TOKEN)
-storage = MemoryStorage()
-dp = Dispatcher(bot, storage=storage)
 
 
 ########################################### <<<MAIN CODE>>>###########################################
@@ -164,8 +148,10 @@ async def chg_nameHandler(message: types.Message, state: FSMContext):
     """
     sqlFormula = "UPDATE user SET name = %s WHERE teleId = %s"
     data = (message.text, message.from_id, )
+    mycursor = db.cursor()
     mycursor.execute(sqlFormula, data)
     db.commit()
+    mycursor.close()
     await message.reply("Okay done, use /myinfo to check")
     await state.finish()
 
@@ -198,7 +184,9 @@ async def chg_roomHandler(message: types.Message, state: FSMContext):
     if check_room_format(s):
         sqlFormula = "UPDATE user SET roomNo = %s WHERE teleId = %s"
         data = (s, message.from_id, )
+        mycursor = db.cursor()
         mycursor.execute(sqlFormula, data)
+        mycursor.close()
         db.commit()
         await message.reply("Okay done, use /myinfo to check")
         await state.finish()
@@ -232,7 +220,9 @@ async def chg_nameHandler1(message: types.Message, state: FSMContext):
     """
     sqlFormula = "UPDATE user SET spotterName = %s WHERE teleId = %s"
     data = (message.text, message.from_id, )
+    mycursor = db.cursor()
     mycursor.execute(sqlFormula, data)
+    mycursor.close()
     db.commit()
     await message.reply("Okay done, use /myinfo to check")
     await state.finish()
@@ -266,7 +256,9 @@ async def chg_roomHandler1(message: types.Message, state: FSMContext):
     if check_room_format(s):
         sqlFormula = "UPDATE user SET spotterRoomNo = %s WHERE teleId = %s"
         data = (s, message.from_id, )
+        mycursor = db.cursor()
         mycursor.execute(sqlFormula, data)
+        mycursor.close()
         db.commit()
         await message.reply("Okay done, use /myinfo to check")
         await state.finish()
